fix: 修正 std::min 模板参数类型不匹配（Linux i64 vs long long）
在 Linux 上 i64 是 long，而 2147483647LL 是 long long，
std::min 要求两个参数类型一致，导致模板推导失败。
将字面量改为 static_cast<i64>(2147483647) 确保类型匹配。

+++ src/common/entity/entities/projectile/AbstractArrowEntity.cpp
@@ -262,7 +262,7 @@ void AbstractArrowEntity::onEntityHit(const RayTraceResult& result)
     if (m_critical) {
         mc::math::Random rng = createRandomFromEntity(*this);
         i32 bonus = rng.nextInt(damage / 2 + 2);
-        damage = static_cast<i32>(std::min(static_cast<i64>(damage) + bonus, 2147483647LL));
+        damage = static_cast<i32>(std::min(static_cast<i64>(damage) + bonus, static_cast<i64>(2147483647)));
     }
 
     // 穿透检查 - 参考 MC 1.16.5 第305-320行
